+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tangleai/core
 
+## 0.24.1
+
+### Patch Changes
+
+- @tangleai/context@0.24.1
+
 ## 0.24.0
 
 ### Patch Changes