+++ packages/documents/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tangleai/documents
 
+## 0.24.1
+
+### Patch Changes
+
+- @tangleai/context@0.24.1
+  - @tangleai/core@0.24.1
+  - @tangleai/models@0.24.1
+
 ## 0.24.0
 
 ### Patch Changes