+++ packages/memory/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tangleai/memory
 
+## 0.24.1
+
+### Patch Changes
+
+- @tangleai/context@0.24.1
+  - @tangleai/core@0.24.1
+
 ## 0.24.0
 
 ### Minor Changes