+++ packages/search/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tangleai/search
 
+## 0.24.1
+
+### Patch Changes
+
+- @tangleai/core@0.24.1
+
 ## 0.24.0
 
 ### Patch Changes